+++ livecheck/__init__.py
@@ -1,4 +1,4 @@
 """livecheck."""
 from __future__ import annotations
 
-__version__ = '0.1.2'
+__version__ = '0.1.3'