+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to
 
 ## [unreleased]
 
+## [0.1.3]
+
 ### Added
 
 - Add configurable Node.js package manager support. #397 Thanks to @inode64.
@@ -114,7 +116,8 @@ and this project adheres to
 
 - When multiple ebuilds are in the same directory, only the latest one will be considered for updating.
 
-[unreleased]: https://github.com/Tatsh/livecheck/compare/v0.1.2...HEAD
+[unreleased]: https://github.com/Tatsh/livecheck/compare/v0.1.3...HEAD
+[0.1.3]: https://github.com/Tatsh/livecheck/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/Tatsh/livecheck/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/Tatsh/livecheck/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/Tatsh/livecheck/compare/v0.0.13...v0.1.0